pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wildhack-transport-planner"
version = "1.0.0"
description = "Forecast-driven transport planning service for WildHack."
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn[standard]>=0.30.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "pyarrow>=16.0.0",
  "torch>=2.2.0",
  "PyYAML>=6.0.0",
  "pydantic>=2.7.0",
  "httpx>=0.27.0",
  "pytest>=8.2.0",
  "lightgbm>=4.3.0",
  "catboost>=1.2.0",
  "jinja2>=3.1.0",
  "python-multipart>=0.0.9",
  "gdown>=5.2.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
